Merger of many components into a single Homepage
