--- src/components/howitworks.tsx
@@ -1,336 +1,0 @@
-import Link from "next/link"
-import { Button } from "@/components/ui/button"
-import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar"
-export function Howitworks() {
-  return (
-    <section className="w-full py-12 md:py-24 lg:py-32" id="howitworks">
-      
-      <div className="container px-4 md:px-6">
-        <div className="flex flex-col items-center justify-center space-y-4 text-center">
-          <div className="space-y-2">
-            <h2 className="text-3xl font-bold tracking-tighter sm:text-5xl">How It Works</h2>
-            <p className="max-w-[900px] text-muted-foreground md:text-xl/relaxed lg:text-base/relaxed xl:text-xl/relaxed">
-              Discover the simple steps to get started with Manoindia.
-            </p>
-          </div>
-        </div>
-        <div className="mx-auto grid max-w-5xl items-start gap-8 sm:grid-cols-2 md:gap-12 lg:grid-cols-5">
-          <div className="grid gap-4">
-            <div className="flex items-center justify-center rounded-full bg-primary p-4 text-primary-foreground">
-              <span className="text-2xl font-bold">1</span>
-            </div>
-            <div className="space-y-2">
-              <h3 className="text-lg font-bold">Sign Up</h3>
-              <p className="text-muted-foreground">Create an account on Manoindia. It's quick and easy!</p>
-            </div>
-          </div>
-          <div className="grid gap-4">
-            <div className="flex items-center justify-center rounded-full bg-primary p-4 text-primary-foreground">
-              <span className="text-2xl font-bold">2</span>
-            </div>
-            <div className="space-y-2">
-              <h3 className="text-lg font-bold">Create Your Profile</h3>
-              <p className="text-muted-foreground">
-                Provide details about your advertising needs or the space you want to rent out.
-              </p>
-            </div>
-          </div>
-          <div className="grid gap-4">
-            <div className="flex items-center justify-center rounded-full bg-primary p-4 text-primary-foreground">
-              <span className="text-2xl font-bold">3</span>
-            </div>
-            <div className="space-y-2">
-              <h3 className="text-lg font-bold">Browse Listings</h3>
-              <p className="text-muted-foreground">
-                Advertisers can browse available spaces, and space providers can view advertiser profiles.
-              </p>
-            </div>
-          </div>
-          <div className="grid gap-4">
-            <div className="flex items-center justify-center rounded-full bg-primary p-4 text-primary-foreground">
-              <span className="text-2xl font-bold">4</span>
-            </div>
-            <div className="space-y-2">
-              <h3 className="text-lg font-bold">Make a Match</h3>
-              <p className="text-muted-foreground">
-                Advertisers and space providers connect based on mutual interests and requirements.
-              </p>
-            </div>
-          </div>
-          <div className="grid gap-4">
-            <div className="flex items-center justify-center rounded-full bg-primary p-4 text-primary-foreground">
-              <span className="text-2xl font-bold">5</span>
-            </div>
-            <div className="space-y-2">
-              <h3 className="text-lg font-bold">Launch Your Campaign</h3>
-              <p className="text-muted-foreground">
-                Advertisers can start their campaigns, and space providers can start earning.
-              </p>
-            </div>
-          </div>
-        </div>
-      </div>
-      <section className="w-full mx-auto mt-10 max-w-4xl p-8 bg-white rounded-lg shadow-md dark:bg-secondary">
-          <div className="grid grid-cols-3 gap-8 mt-6">
-            <div className="flex flex-col items-center text-center">
-              <SearchIcon className="w-12 h-12 text-primary" />
-              <h3 className="mt-4 text-lg font-medium">Search Slots</h3>
-              <p className="mt-2 text-muted-foreground">Browse our available ad slots.</p>
-            </div>
-            <div className="flex flex-col items-center text-center">
-              <CalendarIcon className="w-12 h-12 text-primary" />
-              <h3 className="mt-4 text-lg font-medium">Book and Pay</h3>
-              <p className="mt-2 text-muted-foreground">Secure your ad slot with payment.</p>
-            </div>
-            <div className="flex flex-col items-center text-center">
-              <UploadIcon className="w-12 h-12 text-primary" />
-              <h3 className="mt-4 text-lg font-medium">Upload Ads</h3>
-              <p className="mt-2 text-muted-foreground">Provide your ad creative and assets.</p>
-            </div>
-          </div>
-        </section>
-    </section>
-  )
-}
-
-function CalendarIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M8 2v4" />
-      <path d="M16 2v4" />
-      <rect width="18" height="18" x="3" y="4" rx="2" />
-      <path d="M3 10h18" />
-    </svg>
-  )
-}
-
-
-function MailIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <rect width="20" height="16" x="2" y="4" rx="2" />
-      <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-    </svg>
-  )
-}
-
-
-function MountainIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="m8 3 4 8 5-5 5 15H2L8 3z" />
-    </svg>
-  )
-}
-
-
-function PaletteIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="13.5" cy="6.5" r=".5" fill="currentColor" />
-      <circle cx="17.5" cy="10.5" r=".5" fill="currentColor" />
-      <circle cx="8.5" cy="7.5" r=".5" fill="currentColor" />
-      <circle cx="6.5" cy="12.5" r=".5" fill="currentColor" />
-      <path d="M12 2C6.5 2 2 6.5 2 12s4.5 10 10 10c.926 0 1.648-.746 1.648-1.688 0-.437-.18-.835-.437-1.125-.29-.289-.438-.652-.438-1.125a1.64 1.64 0 0 1 1.668-1.668h1.996c3.051 0 5.555-2.503 5.555-5.554C21.965 6.012 17.461 2 12 2z" />
-    </svg>
-  )
-}
-
-
-function PhoneIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M22 16.92v3a2 2 0 0 1-2.18 2 19.79 19.79 0 0 1-8.63-3.07 19.5 19.5 0 0 1-6-6 19.79 19.79 0 0 1-3.07-8.67A2 2 0 0 1 4.11 2h3a2 2 0 0 1 2 1.72 12.84 12.84 0 0 0 .7 2.81 2 2 0 0 1-.45 2.11L8.09 9.91a16 16 0 0 0 6 6l1.27-1.27a2 2 0 0 1 2.11-.45 12.84 12.84 0 0 0 2.81.7A2 2 0 0 1 22 16.92z" />
-    </svg>
-  )
-}
-
-
-function RocketIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M4.5 16.5c-1.5 1.26-2 5-2 5s3.74-.5 5-2c.71-.84.7-2.13-.09-2.91a2.18 2.18 0 0 0-2.91-.09z" />
-      <path d="m12 15-3-3a22 22 0 0 1 2-3.95A12.88 12.88 0 0 1 22 2c0 2.72-.78 7.5-6 11a22.35 22.35 0 0 1-4 2z" />
-      <path d="M9 12H4s.55-3.03 2-4c1.62-1.08 5 0 5 0" />
-      <path d="M12 15v5s3.03-.55 4-2c1.08-1.62 0-5 0-5" />
-    </svg>
-  )
-}
-
-
-function SearchIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="11" cy="11" r="8" />
-      <path d="m21 21-4.3-4.3" />
-    </svg>
-  )
-}
-
-
-function SlidersVerticalIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <line x1="4" x2="4" y1="21" y2="14" />
-      <line x1="4" x2="4" y1="10" y2="3" />
-      <line x1="12" x2="12" y1="21" y2="12" />
-      <line x1="12" x2="12" y1="8" y2="3" />
-      <line x1="20" x2="20" y1="21" y2="16" />
-      <line x1="20" x2="20" y1="12" y2="3" />
-      <line x1="2" x2="6" y1="14" y2="14" />
-      <line x1="10" x2="14" y1="8" y2="8" />
-      <line x1="18" x2="22" y1="16" y2="16" />
-    </svg>
-  )
-}
-
-
-function StarIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2" />
-    </svg>
-  )
-}
-
-
-function UploadIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
-      <polyline points="17 8 12 3 7 8" />
-      <line x1="12" x2="12" y1="3" y2="15" />
-    </svg>
-  )
-}
-
-
-function XIcon(props:any) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M18 6 6 18" />
-      <path d="m6 6 12 12" />
-    </svg>
-  )
-}
